+++ lang/ar/messages.php
@@ -32,6 +32,8 @@ return [
     "Deleted Item" => "تم الحذف بنجاح",
     "Faild delete item" => "لم يتم حذف العنصر",
     "Item Saved" => "تم الحفظ",
+    "Don't Have any type yet" => 'لا يوجد انواع بعد',
+    "Don't support this action yet" => "هذه العملية غير مدعومة حاليا",
 ];