+++ lang/en/string.php
@@ -143,6 +143,7 @@ return [
     "Add new Collage Information" =>  "Add new Collage Information " ,
     "Edit Collage Information" => "Edit Collage Information",
     "Edit Request Information" => "Edit Request Information",
+    "Edit Request Type" => "Edit request type",
 ];