+++ resources/views/dashboard/admin/requests/type/create.blade.php
@@ -1,0 +1,12 @@
+<x-layouts.dashboard>
+
+    <x-slot:title>
+       create type request
+    </x-slot:title>
+
+    <livewire:form-create-request-type />
+
+
+
+
+</x-layouts.dashboard>