+++ resources/views/dashboard/admin/requests/type/edit.blade.php
@@ -1,0 +1,8 @@
+<x-layouts.dashboard>
+
+    <x-slot:title>
+        type request
+    </x-slot:title>
+    <livewire:form-update-request-type :id="$id" />
+
+</x-layouts.dashboard>