+++ resources/views/dashboard/admin/requests/type/index.blade.php
@@ -1,0 +1,8 @@
+<x-layouts.dashboard>
+
+    <x-slot:title>
+        type request
+    </x-slot:title>
+    <livewire:data-tables.request-type-table />
+
+</x-layouts.dashboard>